sdp::description

Module session

Structs

Origin

Origin defines the structure for the “o=” field which provides the originator of the session plus a session identifier and version number.

RepeatTime

RepeatTime describes the “r=” fields of the session description which represents the intervals and durations for repeated scheduled sessions.

SessionDescription

SessionDescription is a a well-defined format for conveying sufficient information to discover and participate in a multimedia session.

TimeDescription

TimeDescription describes “t=”, “r=” fields of the session description which are used to specify the start and stop times for a session as well as repeat intervals and durations for the scheduled session.

TimeZone

TimeZone defines the structured object for “z=” line which describes repeated sessions scheduling.

Timing

Timing defines the “t=” field’s structured representation for the start and stop times.

Constants

ATTR_KEY_CANDIDATE

Constants for SDP attributes used in JSEP

ATTR_KEY_CONNECTION_SETUP

ATTR_KEY_END_OF_CANDIDATES

ATTR_KEY_EXTMAP_ALLOW_MIXED

ATTR_KEY_EXT_MAP

ATTR_KEY_GROUP

ATTR_KEY_ICELITE

ATTR_KEY_IDENTITY

ATTR_KEY_INACTIVE

ATTR_KEY_MID

ATTR_KEY_MSID

ATTR_KEY_MSID_SEMANTIC

ATTR_KEY_RECV_ONLY

ATTR_KEY_RTCPMUX

ATTR_KEY_RTCPRSIZE

ATTR_KEY_SEND_ONLY

ATTR_KEY_SEND_RECV

ATTR_KEY_SSRC

ATTR_KEY_SSRCGROUP

SEMANTIC_TOKEN_FLOW_IDENTIFICATION

SEMANTIC_TOKEN_FORWARD_ERROR_CORRECTION

SEMANTIC_TOKEN_LIP_SYNCHRONIZATION

Constants for semantic tokens used in JSEP

SEMANTIC_TOKEN_WEBRTC_MEDIA_STREAMS

Type Aliases

EmailAddress

EmailAddress describes a structured representations for the “e=” line which specifies email contact information for the person responsible for the conference.

PhoneNumber

PhoneNumber describes a structured representations for the “p=” line specify phone contact information for the person responsible for the conference.

SessionName

SessionName describes a structured representations for the “s=” field and is the textual session name.

Version

Version describes the value provided by the “v=” field which gives the version of the Session Description Protocol.